ref(alerts): Move anomaly detection code out of subscription processor
