+++ src/sentry/incidents/subscription_processor.py
@@ -12,10 +12,8 @@ from django.db import router, transaction
 from django.utils import timezone
 from sentry_redis_tools.retrying_cluster import RetryingRedisCluster
 from snuba_sdk import Column, Condition, Limit, Op
-from urllib3.exceptions import MaxRetryError, TimeoutError
 
 from sentry import features
-from sentry.conf.server import SEER_ANOMALY_DETECTION_ENDPOINT_URL
 from sentry.incidents.logic import (
     CRITICAL_TRIGGER_LABEL,
     WARNING_TRIGGER_LABEL,
@@ -46,17 +44,8 @@ from sentry.incidents.models.incident import (
 from sentry.incidents.tasks import handle_trigger_action
 from sentry.incidents.utils.types import QuerySubscriptionUpdate
 from sentry.models.project import Project
-from sentry.net.http import connection_from_url
-from sentry.seer.anomaly_detection.types import (
-    AlertInSeer,
-    AnomalyDetectionConfig,
-    AnomalyType,
-    DetectAnomaliesRequest,
-    DetectAnomaliesResponse,
-    TimeSeriesPoint,
-)
-from sentry.seer.anomaly_detection.utils import translate_direction
-from sentry.seer.signed_seer_api import make_signed_seer_api_request
+from sentry.seer.anomaly_detection.get_anomaly_data import get_anomaly_data_from_seer
+from sentry.seer.anomaly_detection.utils import anomaly_has_confidence, has_anomaly
 from sentry.snuba.dataset import Dataset
 from sentry.snuba.entity_subscription import (
     ENTITY_TIME_COLUMNS,
@@ -65,9 +54,8 @@ from sentry.snuba.entity_subscription import (
 )
 from sentry.snuba.models import QuerySubscription
 from sentry.snuba.subscriptions import delete_snuba_subscription
-from sentry.utils import json, metrics, redis
+from sentry.utils import metrics, redis
 from sentry.utils.dates import to_datetime
-from sentry.utils.json import JSONDecodeError
 
 logger = logging.getLogger(__name__)
 REDIS_TTL = int(timedelta(days=7).total_seconds())
@@ -103,11 +91,6 @@ class SubscriptionProcessor:
         AlertRuleThresholdType.ABOVE: (operator.gt, operator.lt),
         AlertRuleThresholdType.BELOW: (operator.lt, operator.gt),
     }
-
-    seer_anomaly_detection_connection_pool = connection_from_url(
-        settings.SEER_ANOMALY_DETECTION_URL,
-        timeout=settings.SEER_ANOMALY_DETECTION_TIMEOUT,
-    )
 
     def __init__(self, subscription: QuerySubscription) -> None:
         self.subscription = subscription
@@ -404,7 +387,7 @@ class SubscriptionProcessor:
                 },
             )
 
-        self.has_anomaly_detection = features.has(
+        has_anomaly_detection = features.has(
             "organizations:anomaly-detection-alerts", self.subscription.project.organization
         ) and features.has(
             "organizations:anomaly-detection-rollout", self.subscription.project.organization
@@ -412,10 +395,15 @@ class SubscriptionProcessor:
 
         potential_anomalies = None
         if (
-            self.has_anomaly_detection
+            has_anomaly_detection
             and self.alert_rule.detection_type == AlertRuleDetectionType.DYNAMIC
         ):
-            potential_anomalies = self.get_anomaly_data_from_seer(aggregation_value)
+            potential_anomalies = get_anomaly_data_from_seer(
+                alert_rule=self.alert_rule,
+                subscription=self.subscription,
+                last_update=self.last_update.timestamp(),
+                aggregation_value=aggregation_value,
+            )
             if potential_anomalies is None:
                 return []
 
@@ -449,7 +437,7 @@ class SubscriptionProcessor:
                     for potential_anomaly in potential_anomalies:
                         # check to see if we have enough data for the dynamic alert rule now
                         if self.alert_rule.status == AlertRuleStatus.NOT_ENOUGH_DATA.value:
-                            if self.anomaly_has_confidence(potential_anomaly):
+                            if anomaly_has_confidence(potential_anomaly):
                                 # NOTE: this means "enabled," and it's the default alert rule status.
                                 # TODO: change these status labels to be less confusing
                                 self.alert_rule.status = AlertRuleStatus.PENDING.value
@@ -458,7 +446,7 @@ class SubscriptionProcessor:
                                 # we don't need to check if the alert should fire if the alert can't fire yet
                                 continue
 
-                        if self.has_anomaly(
+                        if has_anomaly(
                             potential_anomaly, trigger.label
                         ) and not self.check_trigger_matches_status(trigger, TriggerStatus.ACTIVE):
                             metrics.incr(
@@ -474,7 +462,7 @@ class SubscriptionProcessor:
                             self.trigger_alert_counts[trigger.id] = 0
 
                         if (
-                            not self.has_anomaly(potential_anomaly, trigger.label)
+                            not has_anomaly(potential_anomaly, trigger.label)
                             and self.active_incident
                             and self.check_trigger_matches_status(trigger, TriggerStatus.ACTIVE)
                         ):
@@ -540,127 +528,6 @@ class SubscriptionProcessor:
         # this will have no effect, but if someone manages to close a triggered incident
         # before the next one then we might alert twice.
         self.update_alert_rule_stats()
-
-    def has_anomaly(self, anomaly: TimeSeriesPoint, label: str) -> bool:
-        """
-        Helper function to determine whether we care about an anomaly based on the
-        anomaly type and trigger type.
-        """
-        anomaly_type = anomaly.get("anomaly", {}).get("anomaly_type")
-
-        if anomaly_type == AnomalyType.HIGH_CONFIDENCE.value or (
-            label == WARNING_TRIGGER_LABEL and anomaly_type == AnomalyType.LOW_CONFIDENCE.value
-        ):
-            return True
-        return False
-
-    def anomaly_has_confidence(self, anomaly: TimeSeriesPoint) -> bool:
-        """
-        Helper function to determine whether we have the 7+ days of data necessary
-        to detect anomalies/send alerts for dynamic alert rules.
-        """
-        anomaly_type = anomaly.get("anomaly", {}).get("anomaly_type")
-        return anomaly_type != AnomalyType.NO_DATA.value
-
-    def get_anomaly_data_from_seer(
-        self, aggregation_value: float | None
-    ) -> list[TimeSeriesPoint] | None:
-        anomaly_detection_config = AnomalyDetectionConfig(
-            time_period=int(self.alert_rule.snuba_query.time_window / 60),
-            sensitivity=self.alert_rule.sensitivity,
-            direction=translate_direction(self.alert_rule.threshold_type),
-            expected_seasonality=self.alert_rule.seasonality,
-        )
-        context = AlertInSeer(
-            id=self.alert_rule.id,
-            cur_window=TimeSeriesPoint(
-                timestamp=self.last_update.timestamp(), value=aggregation_value
-            ),
-        )
-        detect_anomalies_request = DetectAnomaliesRequest(
-            organization_id=self.subscription.project.organization.id,
-            project_id=self.subscription.project_id,
-            config=anomaly_detection_config,
-            context=context,
-        )
-        extra_data = {
-            "subscription_id": self.subscription.id,
-            "dataset": self.subscription.snuba_query.dataset,
-            "organization_id": self.subscription.project.organization.id,
-            "project_id": self.subscription.project_id,
-            "alert_rule_id": self.alert_rule.id,
-        }
-        try:
-            response = make_signed_seer_api_request(
-                self.seer_anomaly_detection_connection_pool,
-                SEER_ANOMALY_DETECTION_ENDPOINT_URL,
-                json.dumps(detect_anomalies_request).encode("utf-8"),
-            )
-        except (TimeoutError, MaxRetryError):
-            logger.warning(
-                "Timeout error when hitting anomaly detection endpoint", extra=extra_data
-            )
-            return None
-
-        if response.status > 400:
-            logger.error(
-                "Error when hitting Seer detect anomalies endpoint",
-                extra={
-                    "response_data": response.data,
-                    **extra_data,
-                },
-            )
-            return None
-        try:
-            decoded_data = response.data.decode("utf-8")
-        except AttributeError:
-            logger.exception(
-                "Failed to parse Seer anomaly detection response",
-                extra={
-                    "ad_config": anomaly_detection_config,
-                    "context": context,
-                    "response_data": response.data,
-                    "response_code": response.status,
-                },
-            )
-            return None
-
-        try:
-            results: DetectAnomaliesResponse = json.loads(decoded_data)
-        except JSONDecodeError:
-            logger.exception(
-                "Failed to parse Seer anomaly detection response",
-                extra={
-                    "ad_config": anomaly_detection_config,
-                    "context": context,
-                    "response_data": decoded_data,
-                    "response_code": response.status,
-                },
-            )
-            return None
-
-        if not results.get("success"):
-            logger.error(
-                "Error when hitting Seer detect anomalies endpoint",
-                extra={
-                    "error_message": results.get("message", ""),
-                    **extra_data,
-                },
-            )
-            return None
-
-        ts = results.get("timeseries")
-        if not ts:
-            logger.warning(
-                "Seer anomaly detection response returned no potential anomalies",
-                extra={
-                    "ad_config": anomaly_detection_config,
-                    "context": context,
-                    "response_data": results.get("message"),
-                },
-            )
-            return None
-        return ts
 
     def calculate_event_date_from_update_date(self, update_date: datetime) -> datetime:
         """
